requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.1
django-configurations==1.0
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-dotenv==1.3.0
django-extensions==1.6.1
django-filter==0.12.0
django-rest-swagger==0.3.4
django-storages-redux==1.3
djangorestframework==3.3.2
gunicorn==19.4.1
Pillow==3.0.0
psycopg2==2.6.1
PyYAML==3.11
six==1.10.0
sqlparse==0.1.18
static3==0.6.1
Werkzeug==0.11.3
wheel==0.26.0
